Move files for consistency
