+++ theme/base/javascripts/wayf/utility/hasSelectedIdps.js
