+++ theme/openconext/javascripts/wayf/focusSearchBar.js
